// src/store/programs/effects/gql/graphql-types/Subjects.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: Subjects
// ====================================================

export interface Subjects_subjects {
  uid: string;
  title: string;
}

export interface Subjects {
  subjects: Subjects_subjects[];
}